Create README.md
My name is Dionisis and I am 29 years old. I live in Thessaloniki for the last ten years and I would like to work as web developer either as a full time or remote. For the past years I was involved in the sales sector but I figured that it was not something that it filled me. To know me better, I would like to add some hobbies and interests of mine. I am a ski instructor and very passionate with the ski in general as a hobby. I also have some degrees in wine tasting, something else that fills me like as ski. Lastly, I like to read books for Marketing, Management, Business Psychology and Businesse in general.

I was always interested in information technology and i wanted to study Web Development since i was passionate with computers and hardware. During the last year i started reading marketing and sales books emphasized on customers' psychology. In the future i want to focus more on Cyber Security.